Rename transaction manager file and fix imports
